requirements.txt
discord.py>=2.0.0
yt-dlp
faster-whisper
pytube
requests
